Linter: Introduce static analysis
Following on from #90 it was noticed that there was some dead code, so
to improve the codebase a static-analysis tool was added and run. The
subsequent changes are the fallout from running the tool.

+++ internal/httpfile/httpfile.go
@@ -158,7 +158,7 @@ func (f *File) Read(buf []byte) (int, error) {
 		return n, err
 	}
 	if err == nil {
-		if f.reader1 == nil && f.reader0.remaining() < int64(f.readAhead/2) && f.reader0.p1 < f.length {
+		if f.reader1 == nil && f.reader0.remaining() < (f.readAhead/2) && f.reader0.p1 < f.length {
 			// We're well advanced through the current reader,
 			// so kick off a new one.
 			f.reader1 = f.newReaderAhead(f.reader0.p1, f.reader0.p1+f.readAhead)
